# docs/swagger.yaml
definitions:
  api.BaseResp:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 响应数据
      message:
        description: 错误信息
        type: string
    type: object
  api.CreateActivityRequest:
    description: 创建活动请求参数
    properties:
      category:
        description: '@Description 活动类型'
        type: string
      end_at:
        description: '@Description 活动结束时间'
        type: integer
      name:
        description: '@Description 活动名称'
        type: string
      start_at:
        description: '@Description 活动开始时间'
        type: integer
      status:
        description: '@Description 活动状态'
        type: integer
      version:
        description: '@Description 活动版本'
        type: string
    required:
    - category
    - end_at
    - name
    - start_at
    - status
    - version
    type: object
  api.CreateActivityResponse:
    description: 创建活动响应数据
    properties:
      id:
        description: '@Description 活动ID'
        type: integer
    type: object
  api.GetActivityResponse:
    description: 获取活动响应数据
    properties:
      category:
        description: '@Description 活动类型'
        type: string
      end_at:
        description: '@Description 活动结束时间'
        type: integer
      id:
        description: '@Description 活动ID'
        type: integer
      name:
        description: '@Description 活动名称'
        type: string
      start_at:
        description: '@Description 活动开始时间'
        type: integer
      status:
        description: '@Description 活动状态'
        type: integer
      version:
        description: '@Description 活动版本'
        type: string
    type: object
  api.GetGameStatusResponse:
    description: 获取玩法状态响应数据
    properties:
      participation_count:
        description: '@Description 参与次数'
        type: integer
      status:
        description: '@Description 玩法状态'
        type: string
    type: object
  api.GetParticipationResponse:
    description: 获取参与记录响应数据
    properties:
      created_at:
        description: '@Description 参与时间'
        type: integer
      participation_id:
        description: '@Description 参与ID'
        type: integer
      user_id:
        description: '@Description 用户ID'
        type: string
    type: object
  api.GetUserPrizeResponse:
    description: 获取用户奖品响应数据
    properties:
      prizes:
        description: '@Description 奖品列表'
        items:
          $ref: '#/definitions/api.PrizeInfo'
        type: array
    type: object
  api.ParticipateGameReq:
    description: 参与玩法请求参数
    properties:
      activity_id:
        description: '@Description 活动ID'
        type: string
      game_name:
        description: '@Description 玩法名称'
        type: string
      user_id:
        description: '@Description 用户ID'
        type: string
    required:
    - activity_id
    - game_name
    - user_id
    type: object
  api.ParticipateGameResponse:
    description: 参与玩法响应数据
    properties:
      has_prize:
        description: '@Description 是否获得奖品'
        type: boolean
      prize:
        allOf:
        - $ref: '#/definitions/api.PrizeInfo'
        description: '@Description 奖品信息'
      success:
        description: '@Description 是否参与成功'
        type: boolean
    type: object
  api.ParticipateRequest:
    description: 参与活动请求参数
    properties:
      user_id:
        description: '@Description 用户ID'
        type: string
    required:
    - user_id
    type: object
  api.ParticipateResponse:
    description: 参与活动响应数据
    properties:
      participation_id:
        description: '@Description 参与ID'
        type: integer
    type: object
  api.PrizeInfo:
    properties:
      discount_code:
        description: 折扣码（折扣码类型）
        type: string
      price_rule_id:
        description: 价格规则ID（折扣码类型）
        type: integer
      sku:
        description: 商品SKU（商品类型）
        type: string
      title:
        description: 商品标题（商品类型）
        type: string
      type:
        description: 奖品类型
        type: string
    type: object
  api.UpdateActivityRequest:
    description: 更新活动请求参数
    properties:
      category:
        description: '@Description 活动类型'
        type: string
      end_at:
        description: '@Description 活动结束时间'
        type: integer
      name:
        description: '@Description 活动名称'
        type: string
      start_at:
        description: '@Description 活动开始时间'
        type: integer
      status:
        description: '@Description 活动状态'
        type: integer
      version:
        description: '@Description 活动版本'
        type: string
    type: object
  api.UpdateActivityResponse:
    description: 更新活动响应数据
    properties:
      success:
        description: '@Description 是否更新成功'
        type: boolean
    type: object
info:
  contact: {}
paths:
  /activity:
    post:
      consumes:
      - application/json
      description: 创建一个新的活动
      parameters:
      - description: 活动信息
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/api.CreateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.CreateActivityResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 创建活动
      tags:
      - 活动管理
  /activity/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定ID的活动详细信息
      parameters:
      - description: 活动ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.GetActivityResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 获取活动信息
      tags:
      - 活动管理
    put:
      consumes:
      - application/json
      description: 更新指定ID的活动信息
      parameters:
      - description: 活动ID
        in: path
        name: id
        required: true
        type: string
      - description: 活动信息
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/api.UpdateActivityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.UpdateActivityResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 更新活动
      tags:
      - 活动管理
  /activity/{id}/participate:
    post:
      consumes:
      - application/json
      description: 用户参与指定活动
      parameters:
      - description: 活动ID
        in: path
        name: id
        required: true
        type: string
      - description: 参与信息
        in: body
        name: participation
        required: true
        schema:
          $ref: '#/definitions/api.ParticipateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.ParticipateResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 参与活动
      tags:
      - 活动管理
  /activity/{id}/participation:
    get:
      consumes:
      - application/json
      description: 获取用户在指定活动中的参与记录
      parameters:
      - description: 活动ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.GetParticipationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 获取参与记录
      tags:
      - 活动管理
  /game/participate:
    post:
      consumes:
      - application/json
      description: 用户参与指定玩法
      parameters:
      - description: 参与信息
        in: body
        name: participation
        required: true
        schema:
          $ref: '#/definitions/api.ParticipateGameReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.ParticipateGameResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 参与玩法
      tags:
      - 玩法管理
  /game/prize:
    get:
      consumes:
      - application/json
      description: 获取用户在指定玩法中获得的奖品
      parameters:
      - description: 活动ID
        in: query
        name: activity_id
        required: true
        type: string
      - description: 玩法名称
        in: query
        name: game_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.GetUserPrizeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 获取用户奖品
      tags:
      - 玩法管理
  /game/status:
    get:
      consumes:
      - application/json
      description: 获取用户在指定玩法中的状态
      parameters:
      - description: 活动ID
        in: query
        name: activity_id
        required: true
        type: string
      - description: 玩法名称
        in: query
        name: game_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.BaseResp'
            - properties:
                data:
                  $ref: '#/definitions/api.GetGameStatusResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.BaseResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BaseResp'
      summary: 获取玩法状态
      tags:
      - 玩法管理
swagger: "2.0"